Air drawing using Augmented Reality
Requirements:
Processing IDE

Description:
patt.hiro is the marker used here.
.pde is the file extension for a program in Processing IDE. When the
program is run, it takes input from camera. It tracks the position of
marker i.e. hiro in this case. A coloured circle is augmented on this
marker. When the marker is moved in space, the program goes on
augmenting black dots on it. This process appears as if marker is the
brush and when you move this marker, a coloured line is being drawn in
air.